refactor(report): 색상 상수를 reportColors로 분리해 역결합 해소
resumeColors / personalStatementColors가 ReportResumePage /
ReportPersonalStatementPage 컴포넌트에서 export 되어 10여 개 섹션이
페이지 컴포넌트를 import 하던 역결합을 제거.
domain/report/reportColors.ts 도메인 공용 상수로 분리하고 모든 import 경로 정리.
