+++ apps/web/src/domain/report/reportColors.ts
@@ -1,0 +1,27 @@
+// 서류 진단(report) 랜딩 페이지 색상 팔레트.
+// 이전에는 ReportResumePage / ReportPersonalStatementPage 컴포넌트에서 export 하여
+// 여러 섹션이 페이지 컴포넌트를 import 하는 역결합이 있었다. 도메인 공용 상수로 분리.
+
+export const resumeColors = {
+  E8FDF2: '#E8FDF2',
+  B1FFD6: '#B1FFD6',
+  A5FFCF: '#A5FFCF',
+  _4FDA46: '#4FDA46',
+  _2CE282: '#2CE282',
+  _06B259: '#06B259',
+  D8E36C: '#D8E36C',
+  F7FFAB: '#F7FFAB',
+  _14BCFF: '#14BCFF',
+  EEFAFF: '#EEFAFF',
+  _2CDDEA: '#2CDDEA',
+  _11AC5C: '#11AC5C',
+};
+
+export const personalStatementColors = {
+  C34AFF: '#C34AFF',
+  F9EEFF: '#F9EEFF',
+  CA60FF: '#CA60FF',
+  EB6CFF: '#EB6CFF',
+  F3A2FF: '#F3A2FF',
+  FCE9FF: '#FCE9FF',
+};
