+++ apps/web/src/domain/report/sections/ReportExampleSection.tsx
@@ -4,8 +4,10 @@ import NextButton from '@/assets/icons/next-button.svg?react';
 import PrevButton from '@/assets/icons/prev-button.svg?react';
 import { REPORT_EXAMPLE } from '@/data/reportConstant';
 import { REPORT_EXAMPLE_ID } from '@/domain/report/ui/ReportNavigation';
-import { personalStatementColors } from '@/domain/report/ReportPersonalStatementPage';
-import { resumeColors } from '@/domain/report/ReportResumePage';
+import {
+  personalStatementColors,
+  resumeColors,
+} from '@/domain/report/reportColors';
 import { useControlScroll } from '@/hooks/useControlScroll';
 import { ReportExample } from '@/types/interface';
 import { useRef, useState } from 'react';
